made abstract parent class
